+++ tt_metal/fabric/channel_trimming_import.cpp
@@ -7,42 +7,17 @@
 #include <limits>
 #include <stdexcept>
 #include <string>
-#include <vector>
 
 #include <tt-logger/tt-logger.hpp>
 #include <tt_stl/assert.hpp>
 #include <yaml-cpp/yaml.h>
 
+#include "tt_metal/fabric/channel_trimming_io.hpp"
 #include "tt_metal/fabric/fabric_edm_packet_header.hpp"  // NocSendType
 
 namespace tt::tt_fabric {
 
 namespace {
-
-// Parse "chip_N" → N
-ChipId parse_chip_key(const std::string& key) {
-    constexpr auto prefix = std::string_view("chip_");
-    TT_FATAL(
-        key.size() > prefix.size() && key.substr(0, prefix.size()) == prefix,
-        "Invalid chip key in trimming profile: '{}'",
-        key);
-    return static_cast<ChipId>(std::stoi(key.substr(prefix.size())));
-}
-
-// Parse "eth_channel_N" → N
-chan_id_t parse_eth_channel_key(const std::string& key) {
-    constexpr auto prefix = std::string_view("eth_channel_");
-    TT_FATAL(
-        key.size() > prefix.size() && key.substr(0, prefix.size()) == prefix,
-        "Invalid eth channel key in trimming profile: '{}'",
-        key);
-    return static_cast<chan_id_t>(std::stoi(key.substr(prefix.size())));
-}
-
-// Parse hex string like "0x001F" → uint16_t
-uint16_t parse_hex_bitfield(const std::string& str) {
-    return static_cast<uint16_t>(std::stoul(str, nullptr, 16));
-}
 
 // Map NocSendType string name back to its enum value.
 // Returns -1 if the string is not recognized.
@@ -137,20 +112,6 @@ void import_noc_send_types(const YAML::Node& chan_data, ChannelTrimmingOverrides
             }
         }
     }
-}
-
-// Collect all scalar keys from a YAML map node into a vector.
-// Must be done before any operator[] lookups on the map's children, because
-// yaml-cpp's operator[] mutates the underlying node (inserting null entries)
-// which corrupts in-progress iteration.
-std::vector<std::string> collect_map_keys(const YAML::Node& map_node) {
-    std::vector<std::string> keys;
-    for (auto it = map_node.begin(); it != map_node.end(); ++it) {
-        if (it->first.IsScalar()) {
-            keys.push_back(it->first.as<std::string>());
-        }
-    }
-    return keys;
 }
 
 }  // namespace
