+++ tt_metal/fabric/channel_trimming_report.cpp
@@ -14,49 +14,11 @@
 #include <yaml-cpp/yaml.h>
 
 #include "tt_metal/fabric/builder/fabric_builder_config.hpp"
+#include "tt_metal/fabric/channel_trimming_io.hpp"
 
 namespace tt::tt_fabric {
 
 namespace {
-
-// Parse "chip_N" → N
-ChipId parse_chip_key(const std::string& key) {
-    constexpr auto prefix = std::string_view("chip_");
-    TT_FATAL(
-        key.size() > prefix.size() && key.substr(0, prefix.size()) == prefix,
-        "Invalid chip key in trimming profile: '{}'",
-        key);
-    return static_cast<ChipId>(std::stoi(key.substr(prefix.size())));
-}
-
-// Parse "eth_channel_N" → N
-chan_id_t parse_eth_channel_key(const std::string& key) {
-    constexpr auto prefix = std::string_view("eth_channel_");
-    TT_FATAL(
-        key.size() > prefix.size() && key.substr(0, prefix.size()) == prefix,
-        "Invalid eth channel key in trimming profile: '{}'",
-        key);
-    return static_cast<chan_id_t>(std::stoi(key.substr(prefix.size())));
-}
-
-// Parse hex string like "0x001F" → uint16_t
-uint16_t parse_hex_bitfield(const std::string& str) {
-    return static_cast<uint16_t>(std::stoul(str, nullptr, 16));
-}
-
-// Collect all scalar keys from a YAML map node into a vector.
-// Must be done before any operator[] lookups on the map's children, because
-// yaml-cpp's operator[] mutates the underlying node (inserting null entries)
-// which corrupts in-progress iteration.
-std::vector<std::string> collect_map_keys(const YAML::Node& map_node) {
-    std::vector<std::string> keys;
-    for (auto it = map_node.begin(); it != map_node.end(); ++it) {
-        if (it->first.IsScalar()) {
-            keys.push_back(it->first.as<std::string>());
-        }
-    }
-    return keys;
-}
 
 // Returns {expected_sender_channels, expected_receiver_channels} for a router
 // given the fabric topology, VC1 presence, and the router's direction string from the YAML.
